requirements.txt
python-dotenv
pymongo
pyyaml
xxhash